refactor: fsd로 마이그레이션
