--- src/app/page.tsx
@@ -1,11 +1,0 @@
-'use client';
-
-import Map from '@/components/map';
-
-export default function Home() {
-  return (
-    <main className="flex h-screen items-center justify-center gap-2 p-10">
-      <Map />
-    </main>
-  );
-}
