+++ src/entities/resort/model/index.ts
@@ -1,0 +1,2 @@
+export { ResortList } from './constants';
+export { type Spot } from './model';
