+++ src/entities/resort/model/model.d.ts
@@ -1,0 +1,3 @@
+import { ResortList } from './constants';
+
+export type Spot = ExcludeArray<(typeof ResortList)[0]['spots']>;
