+++ src/shared/icons/battery.tsx
