+++ src/shared/icons/cloud.tsx
