+++ src/shared/icons/sun.tsx
