+++ src/shared/icons/video.tsx
