+++ src/shared/icons/wifi.tsx
