+++ src/shared/lib/util.d.ts
@@ -1,0 +1,1 @@
+export type ExcludeArray<T> = T extends Array<infer U> ? U : T;
