+++ src/widgets/header/ui/index.ts
@@ -1,0 +1,2 @@
+export { default as NavBar } from '@/widgets/header/ui/nav-bar';
+export { default as StatusBar } from '@/widgets/header/ui/status-bar';
